+++ .github/workflows/cicd/Package.resolved
@@ -5,8 +5,8 @@
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/vapor/jwt-kit",
       "state" : {
-        "revision" : "87ce13a1df913ba4d51cf00606df7ef24d455571",
-        "version" : "4.7.0"
+        "revision" : "9e929d925434b91857661bcd455d1bd53f00bf22",
+        "version" : "4.13.0"
       }
     },
     {
@@ -28,15 +28,6 @@
       }
     },
     {
-      "identity" : "shellout",
-      "kind" : "remoteSourceControl",
-      "location" : "https://github.com/JohnSundell/ShellOut",
-      "state" : {
-        "revision" : "e1577acf2b6e90086d01a6d5e2b8efdaae033568",
-        "version" : "2.3.0"
-      }
-    },
-    {
       "identity" : "swift-arguments",
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/clayellis/swift-arguments",
@@ -46,12 +37,12 @@
       }
     },
     {
-      "identity" : "swift-ci",
+      "identity" : "swift-cicd",
       "kind" : "remoteSourceControl",
-      "location" : "https://github.com/clayellis/swift-ci",
+      "location" : "https://github.com/swiftcicd/swift-cicd",
       "state" : {
-        "branch" : "action",
-        "revision" : "4d20ce1b9f7c0b16abcea61bd61aa987f40bf69b"
+        "branch" : "main",
+        "revision" : "8d37d8fcd092df0cd199a5977e2b9ab309282f6c"
       }
     },
     {
@@ -59,8 +50,8 @@
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/apple/swift-crypto.git",
       "state" : {
-        "revision" : "92a04c10fc5ce0504f8396aac7392126033e547c",
-        "version" : "2.2.2"
+        "revision" : "33a20e650c33f6d72d822d558333f2085effa3dc",
+        "version" : "2.5.0"
       }
     },
     {
@@ -77,8 +68,8 @@
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/apple/swift-log",
       "state" : {
-        "revision" : "6fe203dc33195667ce1759bf0182975e4653ba1c",
-        "version" : "1.4.4"
+        "revision" : "32e8d724467f8fe623624570367e3d50c5638e46",
+        "version" : "1.5.2"
       }
     },
     {
