+++ Sources/SwiftCICDActions/Git/Push.swift
@@ -1,13 +1,13 @@
 import SwiftCICDCore
 
-public struct GitPush: Action {
-    public func run() async throws -> String {
+struct GitPush: Action {
+    func run() async throws -> String {
         let push = ShellCommand("git push")
         return try await shell(push)
     }
 }
 
-public extension Action {
+public extension Git {
     @discardableResult
     func push() async throws -> String {
         try await run(GitPush())
